Extract uploading and index creating into separate file
